@@ -1,6 +1,6 @@
 /*
  * oerc, alias oer-collector
- * Copyright (C) 2021-2025 emschu[aet]mailbox.org
+ * Copyright (C) 2021-2026 emschu[aet]mailbox.org
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Affero General Public License as
@@ -22,9 +22,10 @@ import {LogEntryResponse} from '../entities';
 import {Subscription} from 'rxjs';
 
 @Component({
-  selector: 'app-logdashboard',
-  templateUrl: './log-dashboard.component.html',
-  styleUrls: ['./log-dashboard.component.scss']
+    selector: 'app-logdashboard',
+    templateUrl: './log-dashboard.component.html',
+    styleUrls: ['./log-dashboard.component.scss'],
+    standalone: false
 })
 export class LogDashboardComponent implements OnInit, OnDestroy {
   logEntries: LogEntryResponse | null = null;
