@@ -1,6 +1,6 @@
 /*
  * oerc, alias oer-collector
- * Copyright (C) 2021-2025 emschu[aet]mailbox.org
+ * Copyright (C) 2021-2026 emschu[aet]mailbox.org
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Affero General Public License as
@@ -25,9 +25,10 @@ import {first} from 'rxjs/operators';
 import dayjs from 'dayjs';
 
 @Component({
-  selector: 'app-recommendation',
-  templateUrl: './recommendation.component.html',
-  styleUrls: ['./recommendation.component.scss']
+    selector: 'app-recommendation',
+    templateUrl: './recommendation.component.html',
+    styleUrls: ['./recommendation.component.scss'],
+    standalone: false
 })
 export class RecommendationComponent extends AbstractReadMoreComponent implements OnInit, OnDestroy {
 
