@@ -1,5 +1,5 @@
 // oerc, alias oer-collector
-// Copyright (C) 2021-2025 emschu[aet]mailbox.org
+// Copyright (C) 2021-2026 emschu[aet]mailbox.org
 //
 // This program is free software: you can redistribute it and/or modify
 // it under the terms of the GNU Affero General Public License as
@@ -36,7 +36,7 @@ import (
 )
 
 var (
-	version       = "0.20.0"
+	version       = "0.21.0"
 	appConf       AppConfig
 	status        Status
 	verboseGlobal = false
